+++ src/main/java/io/github/astrapi69/swing/button/builder/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains model and builder class for generate {@link javax.swing.JButton} objects
+ */
+package io.github.astrapi69.swing.button.builder;