+++ src/main/java/io/github/astrapi69/swing/button/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This is the package for the button icon factory classes
+ */
+package io.github.astrapi69.swing.button;